refactor: footer 파일명 index로 수정
